moved Complaint and Contract back to 'domain' package where they first resided
