+++ app/repositories/Complaint/ComplaintRepository.scala
@@ -5,7 +5,7 @@ import com.websudos.phantom.CassandraTable
 import com.websudos.phantom.dsl._
 import com.websudos.phantom.reactivestreams._
 import conf.DataConnection
-import domain.Complaint.Complaint
+import domain.Complaint
 
 import scala.concurrent.Future
 /**
